requirements.txt
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
datasets>=2.14.0
sentence-transformers>=2.2.0
hdbscan>=0.8.33
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tqdm>=4.65.0
openai>=1.0.0
rouge-score>=0.1.2
bert-score>=0.3.13
python-Levenshtein>=0.21.1
jellyfish>=1.0.0
scipy>=1.11.0
umap-learn>=0.5.4
nltk>=3.8
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0